feat: predefine combine reducers functionality
