+++ src/store/index.tsx
@@ -1,12 +1,13 @@
 import React, { FC } from 'react';
 import { Provider } from 'react-redux';
-import { appReducer } from './appReducer';
-import { RMApi as Api } from '../utils/api';
-import { createStore } from 'redux';
+import { dashboardReducer } from '../modules/Dashboard/dashboardReducer';
+import { createStore, combineReducers } from 'redux';
+import { State } from '../types';
 
-const RMApi = new Api();
-export const initialState = { RMApi, characterPage: [], total: 0 };
-const store = createStore(appReducer, initialState);
+const appReducer = combineReducers<State>({
+  dashboardReducer,
+});
+const store = createStore(appReducer);
 
 export const AppState: FC = ({ children }) => {
   return <Provider store={store}>{children}</Provider>;
